+++ src/main/java/browser/Browser.java
@@ -1,0 +1,18 @@
+package browser;
+import org.openqa.selenium.WebDriver;
+
+public class Browser {
+    private static Browser instance;
+    public WebDriver driver;
+
+    private Browser() {
+        driver = BrowserFactory.getDriver();
+    }
+
+    public static Browser getInstance() {
+        if (instance == null) {
+            instance = new Browser();
+        }
+        return instance;
+    }
+}
