+++ src/test/java/cases/RailcarMarshalingEditTestsPositive.java
@@ -1,0 +1,4 @@
+package cases;
+
+public class RailcarMarshalingEditTestsPositive {
+}
