+++ src/test/java/helpers/LoginData.java
@@ -1,0 +1,14 @@
+package helpers;
+
+public class LoginData {
+    public String login;
+    public String password;
+    public String language;
+
+    public LoginData(String login, String password, String language) {
+        this.login = login;
+        this.password = password;
+        this.language = language;
+    }
+
+}
