--- src/test/java/robotest/base/baseForms/Button.java
@@ -1,13 +1,0 @@
-package robotest.base.baseForms;
-
-import robotest.base.waiting.WebDriverWaits;
-
-public class Button extends BaseElement {
-    public Button(String locator) {
-        super(locator);
-    }
-
-    public String getHref() {
-        return (new WebDriverWaits().explicitWaitOfElement(locator)).getAttribute("href");
-    }
-}
