--- src/test/java/robotest/base/parsers/configManager.java
@@ -1,11 +1,0 @@
-package robotest.base.parsers;
-
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-
-public class configManager {
-
-    DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-    DocumentBuilder builder;
-
-}
