--- src/test/java/robotest/test/cases/RailcarMarshalingEditTestsPositive.java
@@ -1,4 +1,0 @@
-package robotest.test.cases;
-
-public class RailcarMarshalingEditTestsPositive {
-}
