--- src/test/java/robotest/test/data/LoginData.java
@@ -1,14 +1,0 @@
-package robotest.test.data;
-
-public class LoginData {
-    public String login;
-    public String password;
-    public String language;
-
-    public LoginData(String login, String password, String language) {
-        this.login = login;
-        this.password = password;
-        this.language = language;
-    }
-
-}
