--- src/test/java/robotest/test/forms/LoginButton.java
@@ -1,9 +1,0 @@
-package robotest.test.forms;
-
-import robotest.base.baseForms.Button;
-
-public class LoginButton extends Button {
-    public LoginButton(String locator) {
-        super(locator);
-    }
-}
